requirements.txt
anyio==3.7.0
asyncio==3.4.3
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
datatable==1.0.0
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.92.0
fastjinja2templates==0.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.3.3
pyotp==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.20.0
